+++ sdk/servicebus/azure-servicebus/samples/sync_samples/sample_code_servicebus.py
@@ -102,7 +102,7 @@ def example_create_servicebus_receiver_sync():
     queue_name = os.environ['SERVICE_BUS_QUEUE_NAME']
     servicebus_client = ServiceBusClient.from_connection_string(conn_str=servicebus_connection_str)
     with servicebus_client:
-        queue_dlq_receiver = servicebus_client.get_queue_receiver(queue_name=queue_name, sub_queue=ServiceBusSubQueue.DeadLetter)
+        queue_dlq_receiver = servicebus_client.get_queue_receiver(queue_name=queue_name, sub_queue=ServiceBusSubQueue.DEAD_LETTER)
     # [END create_queue_deadletter_receiver_from_sb_client_sync]
 
     # [START create_servicebus_receiver_from_sb_client_sync]
@@ -140,7 +140,7 @@ def example_create_servicebus_receiver_sync():
         subscription_dlq_receiver = servicebus_client.get_subscription_receiver(
             topic_name=topic_name,
             subscription_name=subscription_name,
-            sub_queue=ServiceBusSubQueue.DeadLetter
+            sub_queue=ServiceBusSubQueue.DEAD_LETTER
         )
     # [END create_subscription_deadletter_receiver_from_sb_client_sync]
 
@@ -299,7 +299,7 @@ def example_receive_deadletter_sync():
                     error_description='description for dead lettering'
                 )
 
-        with servicebus_client.get_queue_receiver(queue_name, sub_queue=ServiceBusSubQueue.DeadLetter) as servicebus_deadletter_receiver:
+        with servicebus_client.get_queue_receiver(queue_name, sub_queue=ServiceBusSubQueue.DEAD_LETTER) as servicebus_deadletter_receiver:
             messages = servicebus_deadletter_receiver.receive_messages(max_wait_time=5)
             for message in messages:
                 servicebus_deadletter_receiver.complete_message(message)