+++ sdk/servicebus/azure-servicebus/tests/test_subscriptions.py
@@ -141,7 +141,7 @@ class ServiceBusSubscriptionTests(AzureMgmtTestCase):
                 topic_name=servicebus_topic.name,
                 subscription_name=servicebus_subscription.name,
                 max_wait_time=5,
-                receive_mode=ServiceBusReceiveMode.PeekLock,
+                receive_mode=ServiceBusReceiveMode.PEEK_LOCK,
                 prefetch_count=10
             ) as receiver:
 
@@ -165,7 +165,7 @@ class ServiceBusSubscriptionTests(AzureMgmtTestCase):
                 topic_name=servicebus_topic.name,
                 subscription_name=servicebus_subscription.name,
                 max_wait_time=5,
-                receive_mode=ServiceBusReceiveMode.PeekLock
+                receive_mode=ServiceBusReceiveMode.PEEK_LOCK
             ) as receiver:
                 count = 0
                 for message in receiver:
@@ -177,9 +177,9 @@ class ServiceBusSubscriptionTests(AzureMgmtTestCase):
             with sb_client.get_subscription_receiver(
                 topic_name=servicebus_topic.name,
                 subscription_name=servicebus_subscription.name,
-                sub_queue = ServiceBusSubQueue.DeadLetter,
+                sub_queue = ServiceBusSubQueue.DEAD_LETTER,
                 max_wait_time=5,
-                receive_mode=ServiceBusReceiveMode.PeekLock
+                receive_mode=ServiceBusReceiveMode.PEEK_LOCK
             ) as dl_receiver:
                 count = 0
                 for message in dl_receiver: