+++ sdk/servicebus/azure-servicebus/CHANGELOG.md
@@ -11,6 +11,7 @@
 * `ServiceBusSender` and `ServiceBusReceiver` are no more reusable and will raise `ValueError` when trying to operate on a closed handler.
 * Rename `ReceiveMode` to `ServiceBusReceiveMode` and `SubQueue` to `ServiceBusSubQueue`, and convert their enum values from ints to human-readable strings.
 * An improper `receive_mode` value will now raise `ValueError` instead of `TypeError` in line with supporting extensible enums.
+* Rename enum values `DeadLetter` to `DEAD_LETTER`, `TransferDeadLetter` to `TRANSFER_DEAD_LETTER`, `PeekLock` to `PEEK_LOCK` and `ReceiveAndDelete` to `RECEIVE_AND_DELETE` to conform to sdk guidelines going forward.
 
 ## 7.0.0b8 (2020-11-05)